Can now loot containers for given quantity
I was letting the Container object simply reduce the quantity of an item (container.item.quantity = container.item.quantity - 1).  I've now made it so that the container uses the remove() method of Inventory.  That came with it's own problems.

Previously when an item from removed from an Inventory it was removed completely.  So, container.inventory.remove(item) would completely remove the item regardless of the quantity of said item.  

Now, calling container.inventory.remove(item, amount_to_remove) handles that.  Upon reaching 0 quantity the item is completely removed from the inventory.